+++ Source/BugsnagMetadata.h
@@ -54,6 +54,13 @@
 
 - (void)clearMetadataInSection:(NSString *_Nonnull)section;
 
+/**
+ * Remove a specific value for a specific key in a specific metadata section.
+ * If either section or key do not exist no action is taken.
+ *
+ * @param section The section name
+ * @param key the metadata key
+ */
 - (void)clearMetadataInSection:(NSString *_Nonnull)section
                            key:(NSString *_Nonnull)key;