+++ iOS/BugsnagTests/Swift Tests/BugsnagSwiftTests.swift
@@ -40,7 +40,7 @@ class BugsnagSwiftTests: XCTestCase {
     func testClearMetadataInSectionIsExposedToSwiftCorrectly() {
         try! Bugsnag.start(with: BugsnagConfiguration(DUMMY_APIKEY_32CHAR_1)!)
         // We don't need to check method's functioning, only that we can call it this way
-        Bugsnag.clearMetadata("testSection")
+        Bugsnag.clearMetadata(section: "testSection")
         
         // A double-check that the method signature is as expected
         let metadata = BugsnagMetadata()