+++ CHANGELOG.md
@@ -26,12 +26,16 @@ Bugsnag Notifiers on other platforms.
 * Add a breadcrumb when Bugsnag first starts with the message "Bugsnag loaded"
   [#445](https://github.com/bugsnag/bugsnag-cocoa/pull/445)
   
-* `Bugsnag.addAttribute:value:tab:` is now `Bugsnag.addMetadataToSection::key:value:`
+* `Bugsnag.addAttribute:value:tab:` is now `Bugsnag.addMetadataToSection:key:value:`
   [#454](https://github.com/bugsnag/bugsnag-cocoa/pull/454)
   
 *  `[Bugsnag clearTab:]` is now `[Bugsnag clearMetadataInSection:]` 
      (Swift: `Bugsnag.clearMetadata(_ section)`)
      [#457](https://github.com/bugsnag/bugsnag-cocoa/pull/457)
+     
+* Added `[Bugsnag clearMetadataInSection:withKey:]`
+  (Swift: `Bugsnag.clearMetadata(section:key:)`)
+  [#462](https://github.com/bugsnag/bugsnag-cocoa/pull/462)
 
 * Added `Bugsnag.getMetadata(_ section)`.  The behaviour is: calling with a valid section
   name will return the metadata for that section if it exists, or `nil` if it does not exist.  Other,